On branch main Your branch is up to date with 'origin/main'.
Changes to be committed:
	modified:   blog/__pycache__/models.cpython-311.pyc
	modified:   blog/__pycache__/views.cpython-311.pyc
	modified:   blog/models.py
	modified:   blog/templates/blog/base.html
	new file:   blog/templates/blog/moneyadded.html
	modified:   blog/views.py
	modified:   db.sqlite3
	modified:   django_learning/__pycache__/urls.cpython-311.pyc
	modified:   django_learning/urls.py
	new file:   media/images/Screenshot_2024-06-09_at_19.04.01.png
	new file:   media/profile_pics/Screenshot_2024-06-09_at_19.02.00.png
	modified:   registration/__pycache__/forms.cpython-311.pyc
	modified:   registration/__pycache__/models.cpython-311.pyc
	modified:   registration/__pycache__/views.cpython-311.pyc
	modified:   registration/forms.py
	new file:   registration/migrations/0014_alter_profile_cash.py
	new file:   registration/migrations/__pycache__/0014_alter_profile_cash.cpython-311.pyc
	modified:   registration/models.py

+++ registration/forms.py
@@ -6,7 +6,7 @@ from .models import Profile
 class UserRegisterForm(UserCreationForm):
     email = forms.EmailField(required=False)
     is_seller = forms.BooleanField(required=False)
-    cash = forms.IntegerField()
+    cash = forms.DecimalField(max_digits=10, decimal_places=2)
 
     class Meta:
         model = User
